Add core files of the open-source 10G MAC IP (nfmac10g).
In order to avoid the non-free licensed Xilinx ten_gig_mac core a
compatible ip core is added to the NetFPGA sume repository. Contributions
to extend, fix or improve this core are welcome.

Interface cores in std library are cloned and nfmac10g sources are added on
top of them.